Relative numbers
So, I've just discovered relative numbering for Vim and I love it.
I'll write a blog post later about it. However, it has already be
set as my default line numbering mode.

Anyway, it's still a good idea to provide absolute line numbering
since it's still very useful to know in which line you're at or
to locate a line given its number (stack traces and so are a good
example of these situations). I could type by myself all the
commands for doing this, but since I'm lazy enough to do that, I've
decided to just write a function that swaps relative and absolute
numbering and then I've mapped a call to that function to an unused
key, such as F5, in both Normal and Insert modes.

Now, every time I press F5, line numbering turns absolute if it's
relative, and relative if it's absolute. Just f**king fancy. I only
blame my keyboard now for making me press FN whenever I just want
to type a function key.