Add support for Windows PowerShell
The plugin is wrapped so that it is only downloaded on Windows.
I don't feel useful to have this on Linux and OS X since there
is no PowerShell there.

This conditional plugin downloading is a good experiment that I
might even use in the future to control other plugins such as
YouCompleteMe.