Check for Git, Curl and Vim on install.sh
If the system does not have Git or Curl, the script will refuse to
do its job, since Git is required to clone the repository and Curl
is required to download vim-plug (although wget could be considered
as an alternative).

If Vim is not present on the system, the script will instead provide
a quick help guide on what does he have to run on the command line
once Vim is installed.